Improve game loop - we no longer need to rely on vsync
This also paves the way for interpolated rendering to support high refresh
rates, HOWEVER it turns out that it is impossible to support this feature
without varying the tick rate (which is undesirable).

If we accept that we want to keep a fixed tick rate of 60fps across all systems
(which is important if we want the gameplay to be fully deterministic), then
the only way to support high refresh rates is by rendering players "in between"
frames. However, since players move so unpredictably, it is impossible to know
where they will be in the next frame, and thus it is impossible to know the
correct position to render them until the next frame arrives.